moved AppSettings to Infrastructure
