+++ SuperPanel.App/Controllers/UsersController.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
-using SuperPanel.App.Helpers;
 using SuperPanel.App.Services.Abstract;
 using System.Collections.Generic;
 using System.Threading.Tasks;
