+++ SuperPanel.App/Services/UserService.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.Extensions.Configuration;
 using SuperPanel.App.Data;
 using SuperPanel.App.DistributedServices.Abstract;
-using SuperPanel.App.Helpers;
 using SuperPanel.App.Models;
 using SuperPanel.App.Services.Abstract;
 using System;
