+++ SuperPanel.App/SuperPanel.App.csproj
@@ -26,6 +26,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="Helpers\" />
     <Folder Include="wwwroot\lib\dayjs\" />
     <Folder Include="wwwroot\lib\toastr\" />
   </ItemGroup>
