+++ SuperPanel.App/Views/partials/Pagination.cshtml
@@ -1,4 +1,4 @@
-﻿@using SuperPanel.App.Helpers;
+﻿@using SuperPanel.App.Infrastructure;
 
 <div class="paginator">
     <div class="paginator-options d-none">
